jwt chanining

+++ src/app/Habit/habitController.js
@@ -16,36 +16,42 @@ exports.postHabits = async function(req, res){
     /**
      * Body : emoge, habitName, contents, goodOrBad
      */
-    userIdx = req.params.userIdx;
+    const userIdFromJWT = req.verifiedToken.userIdx;
+    const userId = req.params.userIdx;
 
-    const {habitName, contents, goodOrBad, emoge} = req.body;
+    if (userIdFromJWT != userId) {
+        res.send(errResponse(baseResponse.USER_ID_NOT_MATCH));
+    }
+    else {
+        const {habitName, contents, goodOrBad, emoge} = req.body;
 
-    //빈 값 체크
-    if(!emoge)
-        return res.send(response(baseResponse.HABIT_EMOGE_EMPTY));
-    if(!habitName)
-        return res.send(response(baseResponse.HABIT_NAME_EMPTY));
-    if (!contents)
-        return res.send(response(baseResponse.HABIT_CONTENTS_EMPTY));
+        //빈 값 체크
+        if (!emoge)
+            return res.send(response(baseResponse.HABIT_EMOGE_EMPTY));
+        if (!habitName)
+            return res.send(response(baseResponse.HABIT_NAME_EMPTY));
+        if (!contents)
+            return res.send(response(baseResponse.HABIT_CONTENTS_EMPTY));
 
 
-    //길이 체크
-    if(emoge.length>1)
-        return res.send(response(baseResponse.HABIT_EMOGE_LENGTH));
-    if(habitName.length>20)
-        return res.send(response(baseResponse.HABIT_NAME_LENGTH));
-    if(contents.length>50)
-        return res.send(response(baseResponse.HABIT_CONTENTS_LENGTH));
+        //길이 체크
+        if (emoge.length > 1)
+            return res.send(response(baseResponse.HABIT_EMOGE_LENGTH));
+        if (habitName.length > 20)
+            return res.send(response(baseResponse.HABIT_NAME_LENGTH));
+        if (contents.length > 50)
+            return res.send(response(baseResponse.HABIT_CONTENTS_LENGTH));
 
-    const habitResponse = await habitService.createHabit(
-        userIdx,
-        habitName,
-        contents,
-        goodOrBad,
-        emoge
-    );
+        const habitResponse = await habitService.createHabit(
+            userId,
+            habitName,
+            contents,
+            goodOrBad,
+            emoge
+        );
 
-    return res.send(habitResponse);
+        return res.send(habitResponse);
+    }
 }
 
 /**
@@ -57,13 +63,18 @@ exports.postHabits = async function(req, res){
 exports.getHabits = async function(req, res){
 
     const userId = req.params.userIdx;
-    //습관 전체 조회
-    const habitListResult = await habitProvider.retrieveHabitList(userId);
-    if(!habitListResult)
-        return res.send(response(baseResponse.HABIT_CONTENT_NULL));
+    const userIdFromJWT = req.verifiedToken.userIdx;
 
-    return res.send(response(baseResponse.SUCCESS,habitListResult));
+    if (userIdFromJWT != userId) {
+        res.send(errResponse(baseResponse.USER_ID_NOT_MATCH));
+    }else {
+        //습관 전체 조회
+        const habitListResult = await habitProvider.retrieveHabitList(userId);
+        if (!habitListResult)
+            return res.send(response(baseResponse.HABIT_CONTENT_NULL));
 
+        return res.send(response(baseResponse.SUCCESS, habitListResult));
+    }
 }
 
 /**
@@ -76,14 +87,18 @@ exports.getHabitById = async function(req,res){
 
     const habitId = req.params.habitIdx;
     const userId = req.params.userIdx;
+    const userIdFromJWT = req.verifiedToken.userIdx;
+    if (userIdFromJWT != userId) {
+        res.send(errResponse(baseResponse.USER_ID_NOT_MATCH));
+    }else {
 
-    const habitByHabitId = await habitProvider.retrieveHabit(userId,habitId);
-    if(!habitByHabitId)
-        return res.send(response(baseResponse.HABIT_CONTENT_NULL));
+        const habitByHabitId = await habitProvider.retrieveHabit(userId, habitId);
+        if (!habitByHabitId)
+            return res.send(response(baseResponse.HABIT_CONTENT_NULL));
 
-    return res.send(response(baseResponse.SUCCESS, habitByHabitId));
+        return res.send(response(baseResponse.SUCCESS, habitByHabitId));
 
-
+    }
 }
 
 /**
@@ -96,26 +111,31 @@ exports.patchHabit = async function (req, res){
 
     const userId = req.params.userIdx;
     const habitId = req.params.habitIdx;
-    const {habitName, contents, emoge} = req.body;
-    //빈 값 체크
-    if(!emoge)
-        return res.send(response(baseResponse.HABIT_EMOGE_EMPTY));
-    if(!habitName)
-        return res.send(response(baseResponse.HABIT_NAME_EMPTY));
-    if (!contents)
-        return res.send(response(baseResponse.HABIT_CONTENTS_EMPTY));
+    const userIdFromJWT = req.verifiedToken.userIdx;
+    if (userIdFromJWT != userId) {
+        res.send(errResponse(baseResponse.USER_ID_NOT_MATCH));
+    }else {
+        const {habitName, contents, emoge} = req.body;
+        //빈 값 체크
+        if (!emoge)
+            return res.send(response(baseResponse.HABIT_EMOGE_EMPTY));
+        if (!habitName)
+            return res.send(response(baseResponse.HABIT_NAME_EMPTY));
+        if (!contents)
+            return res.send(response(baseResponse.HABIT_CONTENTS_EMPTY));
 
 
-    //길이 체크
-    if(emoge.length>1)
-        return res.send(response(baseResponse.HABIT_EMOGE_LENGTH));
-    if(habitName.length>20)
-        return res.send(response(baseResponse.HABIT_NAME_LENGTH));
-    if(contents.length>50)
-        return res.send(response(baseResponse.HABIT_CONTENTS_LENGTH));
+        //길이 체크
+        if (emoge.length > 1)
+            return res.send(response(baseResponse.HABIT_EMOGE_LENGTH));
+        if (habitName.length > 20)
+            return res.send(response(baseResponse.HABIT_NAME_LENGTH));
+        if (contents.length > 50)
+            return res.send(response(baseResponse.HABIT_CONTENTS_LENGTH));
 
-    const editHabitInfo = await habitService.editHabit(userId,habitId,habitName, contents, emoge);
-    return res.send(editHabitInfo);
+        const editHabitInfo = await habitService.editHabit(userId, habitId, habitName, contents, emoge);
+        return res.send(editHabitInfo);
+    }
 }
 /**
  * API No. 5
@@ -127,9 +147,15 @@ exports.deleteHabit = async function (req, res){
 
     const userId = req.params.userIdx;
     const habitId = req.params.habitIdx;
+    const userIdFromJWT = req.verifiedToken.userIdx;
 
-    const deleteHabit = await habitService.deleteHabit(userId,habitId);
-    return res.send(deleteHabit);
+    if (userIdFromJWT != userId) {
+        res.send(errResponse(baseResponse.USER_ID_NOT_MATCH));
+    }else {
+
+        const deleteHabit = await habitService.deleteHabit(userId, habitId);
+        return res.send(deleteHabit);
+    }
 }
 
 /**
@@ -143,27 +169,33 @@ exports.checkHabit = async function (req, res){
     const check = req.body;
     const userId = req.params.userIdx;
     const habitId = req.params.habitIdx;
+    const userIdFromJWT = req.verifiedToken.userIdx;
 
-    const habitDay = await habitProvider.getHabitDay(userId,habitId);
-    const habitLife = await habitProvider.getHabitLife(userId,habitId);
+    if (userIdFromJWT != userId) {
+        res.send(errResponse(baseResponse.USER_ID_NOT_MATCH));
+    }else {
 
-    if(habitDay[0][0].habitDay<=0){ //습관 남은 일 수가 0일 때 처리하는 logic
-        const achieveHabit = await habitService.achieveHabit(userId,habitId);
-        return res.send(response(baseResponse.HABIT_ACHIEVEMENT_SUCCESS));
-    }
+        const habitDay = await habitProvider.getHabitDay(userId, habitId);
+        const habitLife = await habitProvider.getHabitLife(userId, habitId);
 
-    if(habitLife[0][0].life<=0){// life가 0이 되어 습관이 삭제되는 logic
-        const deleteHabit =await habitService.deleteHabit(userId,habitId);
-        return res.send(response(baseResponse.HABIT_DELETE_SUCCESS));
-    }
+        if (habitDay[0][0].habitDay <= 0) { //습관 남은 일 수가 0일 때 처리하는 logic
+            const achieveHabit = await habitService.achieveHabit(userId, habitId);
+            return res.send(response(baseResponse.HABIT_ACHIEVEMENT_SUCCESS));
+        }
+
+        if (habitLife[0][0].life <= 0) {// life가 0이 되어 습관이 삭제되는 logic
+            const deleteHabit = await habitService.deleteHabit(userId, habitId);
+            return res.send(response(baseResponse.HABIT_DELETE_SUCCESS));
+        }
 
 
-    if(check.check){
-        const checkHabit = await habitService.checkHabit(userId,habitId);
-        return res.send(checkHabit);
-    }else{
-        const noCheckHabit = await habitService.noCheckHabit(userId,habitId);
-        return res.send(noCheckHabit);
+        if (check.check) {
+            const checkHabit = await habitService.checkHabit(userId, habitId);
+            return res.send(checkHabit);
+        } else {
+            const noCheckHabit = await habitService.noCheckHabit(userId, habitId);
+            return res.send(noCheckHabit);
+        }
     }
 
 }
